+++ spec/dry/types/constrained_spec.rb
@@ -111,7 +111,7 @@ RSpec.describe Dry::Types::Constrained do
     subject(:type) do
       Dry::Types['strict.array']
         .constrained(size: 3)
-        .member(Dry::Types['coercible.string'])
+        .of(Dry::Types['coercible.string'])
     end
 
     it_behaves_like Dry::Types::Definition